pyproject.toml
[tool.poetry]
name = "referral-system"
version = "0.1.0"
description = ""
authors = ["denisbelenko <denisdel123@>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
pillow = "^10.3.0"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
flake8 = "^7.1.0"
django-cors-headers = "^4.4.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.39"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
